django_lesson_5/requirements.txt
Django==2.2.2
ipython
django-ckeditor
psycopg2==2.8.3
django-modeltranslation==0.13.2